Remove stale AI workbench routes
