+++ src/App.tsx
@@ -21,8 +21,6 @@ import APICases from "./pages/cases/APICases";
 import UICases from "./pages/cases/UICases";
 import PerformanceCases from "./pages/cases/PerformanceCases";
 import AICases from "./pages/cases/AICases";
-import AICaseCreate from "./pages/cases/AICaseCreate";
-import AICaseHistory from "./pages/cases/AICaseHistory";
 import Reports from "./pages/reports/Reports";
 import ReportDetail from "./pages/reports/ReportDetail";
 import SystemSettings from "./pages/settings/SystemSettings";
@@ -167,23 +165,9 @@ function Router() {
             </Layout>
           </ProtectedRoute>
         </Route>
-        <Route path="/cases/ai-create">
-          <ProtectedRoute>
-            <Layout>
-              <AICaseCreate />
-            </Layout>
-          </ProtectedRoute>
-        </Route>
         {/* /cases/ai 路由由 KeepAliveAiCases 接管，Switch 中仅保留空占位避免 404 */}
         <Route path="/cases/ai">
           {null}
-        </Route>
-        <Route path="/cases/ai-history">
-          <ProtectedRoute>
-            <Layout>
-              <AICaseHistory />
-            </Layout>
-          </ProtectedRoute>
         </Route>
 
         <Route path="/tasks">
