+++ src/components/Sidebar.tsx
@@ -18,7 +18,6 @@ import {
   Monitor,
   Gauge,
   BrainCircuit,
-  History,
   PanelLeftClose,
   PanelLeftOpen
 } from "lucide-react";
@@ -54,8 +53,7 @@ const navItems: NavItem[] = [
     icon: <BrainCircuit className="h-5 w-5" />,
     label: "AI 工作台",
     children: [
-      { label: "工作台首页", href: "/cases/ai-create", icon: <BrainCircuit className="h-4 w-4" /> },
-      { label: "全部记录", href: "/cases/ai-history", icon: <History className="h-4 w-4" /> },
+      { label: "工作台首页", href: "/cases/ai", icon: <BrainCircuit className="h-4 w-4" /> },
     ],
   },
   { icon: <Boxes className="h-5 w-5" />, label: "任务管理", href: "/tasks" },
