+++ src/pages/cases/components/NewRequirementSheet.tsx
@@ -114,7 +114,7 @@ export function NewRequirementSheet({ open, onOpenChange }: NewRequirementSheetP
       params.set('initReq', requirementText.trim());
       setLocation(`/cases/ai?${params.toString()}`);
     } catch (error) {
-      console.error('[AICaseCreate] failed to navigate', error);
+      console.error('[AIWorkbench] failed to navigate', error);
       toast.error('跳转失败，请重试');
     } finally {
       setIsSubmitting(false);
